test: refactor AttachTest, add detach tests
